move interface tests to their own file + black formatting
